+++ test/doz.single.function.component.js
@@ -33,5 +33,38 @@ describe('Doz.single.function.component', function () {
                 done();
         });
 
+        it('should be ok, simulate babel transform class to function', function (done) {
+
+            document.body.innerHTML = `<div id="app"></div>`;
+
+            class A {}
+
+            const instance = new A();
+
+            Doz.component('salutation-card', function () {
+                if (!(instance instanceof A)) { throw new TypeError("Cannot call a class as a function"); }
+                return `
+                    <div>${this.props.title} ${this.props.name}</div>
+                `
+            });
+
+            try {
+
+                new Doz({
+                    root: '#app',
+                    template: `
+                    <div>
+                    <salutation-card
+                        title="MR."
+                        name="Doz" />
+                        </div>
+                `
+                });
+            } catch (e) {
+                done();
+            }
+
+        });
+
     });
 });